monitor-batch/requirements.txt
gunicorn==20.0.4
APScheduler==3.6.3
PyMySQL==0.9.3
PyYAML==5.1.2
